refactor(chat): better folder structure
