+++ frontend/app/chat/components/ChatsList/ChatsListItem.tsx
@@ -1,0 +1,9 @@
+import { FC } from "react";
+
+interface ChatsListItemProps {}
+
+const ChatsListItem: FC<ChatsListItemProps> = ({}) => {
+  return <div>ChatsListItem</div>;
+};
+
+export default ChatsListItem;
