+++ frontend/app/chat/components/index.ts
@@ -1,4 +1,4 @@
-export * from "./ChatInput";
-export * from "./ChatMessage";
 export * from "./ChatMessages";
+export * from "./ChatMessages/ChatInput";
+export * from "./ChatMessages/ChatMessage";
 export * from "./ChatsList";
