+++ frontend/app/chat/layout.tsx
@@ -1,8 +1,6 @@
 "use client";
-import { useRouter } from "next/navigation";
-import { FC, ReactNode, useState } from "react";
+import { FC, ReactNode } from "react";
 import { ChatsList } from "./components";
-import { Chat } from "./types";
 
 interface LayoutProps {
   children?: ReactNode;
@@ -11,9 +9,7 @@ interface LayoutProps {
 const Layout: FC<LayoutProps> = ({ children }) => {
   return (
     <div className="flex h-screen pt-20 overflow-hidden">
-      <aside className="w-1/5 h-full border-r overflow-auto">
-        <ChatsList />
-      </aside>
+      <ChatsList />
       {children}
     </div>
   );
